Rename flavours/glitch/features/drawer to flavours/glitch/features/compose
