+++ app/javascript/flavours/glitch/features/compose/header/index.js
