+++ app/javascript/flavours/glitch/features/compose/search/index.js
