+++ app/javascript/flavours/glitch/features/compose/search/popout/index.js
