+++ app/javascript/flavours/glitch/theme.yml
@@ -11,7 +11,7 @@ pack:
   home:
     filename: packs/home.js
     preload:
-    - flavours/glitch/async/drawer
+    - flavours/glitch/async/compose
     - flavours/glitch/async/getting_started
     - flavours/glitch/async/home_timeline
     - flavours/glitch/async/notifications
